Added the ability to set the frequency of a node appearing.
Every node has a spawn range. The higher the range, the more likely it will spawn.
The GameManger generates a random number based on the total of the ranges. Then it uses that random number to pick a prefab.

GameManager also creates a list of statistical spawn rates to help keep track of things.

I've set some ranges already, but feel free to change them.